requirements.txt
streamlit
langchain
langchain-google-genai
requests
python-dotenv
pandas
plotly
